requirements.txt
authlib
python-dotenv==1.0.0
google-cloud-firestore==2.11.1
slack-bolt==1.16.1
google-auth>=2.23.0
firebase-admin>=6.0.0
requests>=2.31.0
flask==2.0.1
google-cloud-dialogflow==2.19.1


